todo funcional
todo funciona no mover nada del mundo

+++ src/Mundo/Moves.java
@@ -3,6 +3,6 @@ package Mundo;
 public interface Moves {
 	
 	void move();
-	void changeDirection(int x,int y);
+	void changeDirection(double x,double y);
 
 }
